+++ app/src/main/java/dev/vitalcc/stukay/StukayApplication.kt
@@ -1,0 +1,5 @@
+package dev.vitalcc.stukay
+
+import android.app.Application
+
+class StukayApplication : Application()