+++ core/design/src/main/java/dev/vitalcc/stukay/core/design/theme/Color.kt
@@ -1,0 +1,13 @@
+package dev.vitalcc.stukay.core.design.theme
+
+import androidx.compose.ui.graphics.Color
+
+val StukayAccent = Color(0xFF146C5A)
+val StukayAccentContainer = Color(0xFFD4F4E9)
+val StukaySignal = Color(0xFFB25B18)
+val StukaySignalContainer = Color(0xFFFFDFC8)
+val StukayThreadSurface = Color(0xFFF5F7F7)
+val StukayThreadSurfaceDark = Color(0xFF101818)
+val StukayStatusOk = Color(0xFF1B7F4A)
+val StukayStatusWarn = Color(0xFFB26400)
+val StukayStatusError = Color(0xFFB3261E)