+++ core/design/src/main/java/dev/vitalcc/stukay/core/design/theme/Shape.kt
@@ -1,0 +1,13 @@
+package dev.vitalcc.stukay.core.design.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Shapes
+import androidx.compose.ui.unit.dp
+
+val StukayShapes = Shapes(
+    extraSmall = RoundedCornerShape(12.dp),
+    small = RoundedCornerShape(18.dp),
+    medium = RoundedCornerShape(24.dp),
+    large = RoundedCornerShape(30.dp),
+    extraLarge = RoundedCornerShape(36.dp),
+)